+++ framework/validators/ImageValidator.php
@@ -166,7 +166,7 @@ class ImageValidator extends FileValidator
     {
         ValidationAsset::register($view);
         $options = $this->getClientOptions($object, $attribute);
-        return 'yii.validation.image(attribute, messages, ' . json_encode($options) . ', deferred);';
+        return 'yii.validation.image(attribute, messages, ' . json_encode($options, JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE) . ', deferred);';
     }
 
     /**