+++ framework/validators/RequiredValidator.php
@@ -107,6 +107,6 @@ class RequiredValidator extends Validator
 
         ValidationAsset::register($view);
 
-        return 'yii.validation.required(value, messages, ' . json_encode($options) . ');';
+        return 'yii.validation.required(value, messages, ' . json_encode($options, JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE) . ');';
     }
 }